requirements.txt
streamlit
duckdb
google-generativeai
pandas
python-dotenv
requests
numpy